requirements.txt
# Core dependencies
aiohttp>=3.8.5
aiomqtt>=1.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# OpenAI API
openai>=1.0.0

# Image processing
Pillow>=10.0.0

# Telegram integration
python-telegram-bot>=20.4

# Logging and monitoring
structlog>=23.1.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development tools (optional)
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
flake8>=6.1.0
